Further refactoring of benefit checks.
