+++ gettsim/benefits/benefit_checks/benefit_checks.py
@@ -1,0 +1,79 @@
+def sum_wohngeld_m_arbeitsl_geld_2_eink(
+    sum_basis_arbeitsl_geld_2_eink, _wohngeld_basis_hh_vermögens_check
+):
+    return sum_basis_arbeitsl_geld_2_eink + _wohngeld_basis_hh_vermögens_check
+
+
+def sum_kinderzuschlag_arbeitsl_geld_2_eink(
+    sum_basis_arbeitsl_geld_2_eink, _kinderzuschlag_m_vermögens_check
+):
+    return sum_basis_arbeitsl_geld_2_eink + _kinderzuschlag_m_vermögens_check
+
+
+def sum_wohngeld_m_kinderzuschlag_arbeitsl_geld_2_eink(
+    sum_basis_arbeitsl_geld_2_eink,
+    _wohngeld_basis_hh_vermögens_check,
+    _kinderzuschlag_m_vermögens_check,
+):
+    return (
+        sum_basis_arbeitsl_geld_2_eink
+        + _wohngeld_basis_hh_vermögens_check
+        + _kinderzuschlag_m_vermögens_check
+    )
+
+
+def _subtract_and_clip_at_zero(minuend, subtrahend):
+    return (minuend - subtrahend).clip(lower=0)
+
+
+def arbeitsl_geld_2_m_wohngeld_m_kinderzuschlag(
+    _regelbedarf_m_vermögens_check, sum_wohngeld_m_kinderzuschlag_arbeitsl_geld_2_eink
+):
+    return (
+        _regelbedarf_m_vermögens_check
+        - sum_wohngeld_m_kinderzuschlag_arbeitsl_geld_2_eink
+    ).clip(lower=0)
+
+
+def arbeitsl_geld_2_m_kinderzuschlag(
+    _regelbedarf_m_vermögens_check, sum_kinderzuschlag_arbeitsl_geld_2_eink
+):
+    return (
+        _regelbedarf_m_vermögens_check - sum_kinderzuschlag_arbeitsl_geld_2_eink
+    ).clip(lower=0)
+
+
+def arbeitsl_geld_2_m_wohngeld_m(
+    _regelbedarf_m_vermögens_check, sum_wohngeld_m_arbeitsl_geld_2_eink
+):
+    return (_regelbedarf_m_vermögens_check - sum_wohngeld_m_arbeitsl_geld_2_eink).clip(
+        lower=0
+    )
+
+
+def arbeitsl_geld_2_m_basis(
+    _regelbedarf_m_vermögens_check, sum_basis_arbeitsl_geld_2_eink
+):
+    return (_regelbedarf_m_vermögens_check - sum_basis_arbeitsl_geld_2_eink).clip(
+        lower=0
+    )
+
+
+def anz_rentner_per_hh(hh_id, rentner):
+    return rentner.groupby(hh_id).transform("sum")
+
+
+def wohngeld_m_vorrang(arbeitsl_geld_2_m_wohngeld_m, arbeitsl_geld_2_m_basis):
+    return (arbeitsl_geld_2_m_wohngeld_m == 0) & (arbeitsl_geld_2_m_basis > 0)
+
+
+def kinderzuschlag_vorrang(arbeitsl_geld_2_m_kinderzuschlag, arbeitsl_geld_2_m_basis):
+    return (arbeitsl_geld_2_m_kinderzuschlag == 0) & (arbeitsl_geld_2_m_basis > 0)
+
+
+def wohngeld_m_kinderzuschlag_vorrang(
+    arbeitsl_geld_2_m_wohngeld_m_kinderzuschlag, arbeitsl_geld_2_m_basis
+):
+    return (arbeitsl_geld_2_m_wohngeld_m_kinderzuschlag == 0) & (
+        arbeitsl_geld_2_m_basis > 0
+    )
